Add optional extra yaargh[import-argh] to proxy 'import argh'
This will install the `yaargh-dummy-argh` package, which installs
a dummy `argh` module which replaces itself with yaargh on import.

This allows unmodified programs to use yaargh.

This isn't something we want to do by default as it's quite invasive and would
prevent users from using both libraries at once. However, it is a good way to
shake out issues in yaargh by replacing argh system-wide.