dummy-argh: Fix bug where submodules were imported twice
Resulting in eg. yaargh.exceptions is not argh.exceptions, it's two instances
of the same file. This is because the name does not exist in sys.modules.

So our fix is to ensure the name exists in sys.modules by also doing this
when argh is imported.

https://stackoverflow.com/questions/24322927/python-how-to-alias-module-name-rename-with-preserving-backward-compatibility
is a good resource for further discussion.